PatrolandAttack script + updated AttackBar script
Updated with PatrolandAttack which should update the character animation. There is a public enum file, CharacterStateEnum.cs, which is accessible to AttackBar and PatrolandAttack. In Attackbar, we set the state. When attack bar is full, he is "Attack" state, when finished, back to "Patrol". Idle is not actually used, just included.

PatrolandAttack checks for these states then either sets triggers attached to an animation transition in the Animator tree for our character which this script is attached to, or directly moves via transform like with MoveToMiddle which sets the character to middle in every state change and HandlePatrolMovement which guides the flow between walk left and wlak right with "Switchdirection" which switches between bools which also control the animation flow between walk left and walk right. Currently, there is an issue with animation stopping after moving to middle, suspect a conflict between Movetomiddle and the animation flow.